circ comp utils and misc
