+++ mrmustard/lab_dev/circuit_components_utils/__init__.py
@@ -1,0 +1,22 @@
+# Copyright 2021 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+A set of components that do not correspond to physical elements of a circuit, but can be used to
+perform useful mathematical calculations.
+"""
+
+from .b_to_ps import *
+from .b_to_q import *
+from .trace_out import *
